Finish moving across things
