--- cfspopcon/deprecated_formulas/radiated_power/__init__.py
@@ -1,17 +1,0 @@
-"""Calculate the radiated power due to fuel and impurity species."""
-from .inherent import calc_bremsstrahlung_radiation, calc_synchrotron_radiation
-from .mavrin_coronal import calc_impurity_radiated_power_mavrin_coronal
-from .mavrin_noncoronal import calc_impurity_radiated_power_mavrin_noncoronal
-from .post_and_jensen import calc_impurity_radiated_power_post_and_jensen
-from .radas import calc_impurity_radiated_power_radas
-from .radiated_power import calc_impurity_radiated_power
-
-__all__ = [
-    "calc_bremsstrahlung_radiation",
-    "calc_synchrotron_radiation",
-    "calc_impurity_radiated_power_mavrin_coronal",
-    "calc_impurity_radiated_power_mavrin_noncoronal",
-    "calc_impurity_radiated_power_post_and_jensen",
-    "calc_impurity_radiated_power_radas",
-    "calc_impurity_radiated_power",
-]
