+++ cfspopcon/formulas/radiated_power/synchrotron.py
@@ -93,4 +93,4 @@ def calc_synchrotron_radiation(
     P_sync_r = 6.25e-3 * ne20 * electron_temp_profile * magnetic_field_on_axis**2 * Phi
     P_sync = integrate_profile_over_volume.unitless_func(P_sync_r, rho, plasma_volume)
 
-    return P_sync
+    return P_sync  # type:ignore[no-any-return]
